refactor: writeChatCompletionChunk and writeChatCompletionEnd
