+++ CopilotKit/packages/backend/src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./utils";
 export * from "./openai";
-export * from "./openai-stream-interceptor";
